♻️ refactor(utils): move send_off_sse function to dedicated transport module
- relocate send_off_sse function from utils.py to transport.py for better modularity
- remove redundant imports from utils.py after function relocation
